requirements.txt
openai
python-dotenv
streamlit
faiss-cpu
numpy
PyPDF2
fpdf2
